requirements.txt
streamlit
nltk
numpy
networkx
scikit-learn
matplotlib
feedparser
requests
beautifulsoup4
readability-lxml
snowballstemmer
pyvis
wordcloud
rouge
pypdf
python-docx
pandas
openpyxl